Included Auctions CRUD

+++ BidWheels/Models/Engine.cs
@@ -6,7 +6,7 @@
 		public string Name { get; set; }
 		public string Type { get; set; }
 		public int HorsePower { get; set; }
-		public double Volume { get; set; }
+		public int Volume { get; set; }
 		public ICollection<Car>? Cars { get; set; }
 	}
 }
